[inductor] Consolidate FlyDSL FlexAttention helpers
