+++ torch/_inductor/kernel/vendored_templates/flydsl/kernels/flex_attn_bwd_gfx950.py
@@ -39,11 +39,15 @@ import flydsl.compiler as flyc
 import flydsl.expr as fx
 from flydsl.expr import const_expr
 
-from .flex_attn_mask import is_causal_document_mask_program
-from .flex_attn_utils import load_scalar, make_global_view, store_scalar
+from .flex_attn_utils import (
+    fast_exp2,
+    is_causal_document_mask_program,
+    load_scalar,
+    make_global_view,
+    store_scalar,
+)
 from .flex_bwd_dkdv import make_dkdv_mfma32_body
 from .flex_bwd_dq import make_dq_mfma32_body
-from .flex_bwd_utils import fast_exp2
 
 _LOG2E = 1.4426950408889634
 _BATCHED_CAUSAL_DOCUMENT_MASK_PROGRAM = (
