--- torch/_inductor/kernel/vendored_templates/flydsl/kernels/flex_attn_mask.py
@@ -1,94 +1,0 @@
-# mypy: allow-untyped-defs
-
-import flydsl.expr as fx
-from flydsl.expr import range_constexpr
-
-
-_CAUSAL_DOCUMENT_MASK_PROGRAM = (
-    ("const_bool", True),
-    ("ge", 2, 3),
-    ("and", 4, 5),
-    ("load_i32", 0, (2,)),
-    ("load_i32", 1, (7,)),
-    ("ge", 3, 8),
-    ("and", 6, 9),
-)
-
-
-def is_causal_document_mask_program(
-    mask_program,
-    mask_program_output,
-    mask_buffer_strides,
-):
-    return (
-        tuple(mask_program) == _CAUSAL_DOCUMENT_MASK_PROGRAM
-        and int(mask_program_output) == 10
-        and len(mask_buffer_strides) == 2
-        and all(len(strides) == 1 for strides in mask_buffer_strides)
-    )
-
-
-def evaluate_mask_program(
-    *,
-    mask_program,
-    mask_program_output,
-    mask_buffer_strides,
-    mask_buffers,
-    load_i32,
-    batch,
-    head,
-    q_pos,
-    kv_pos,
-):
-    values = [fx.Int32(batch), fx.Int32(head), q_pos, kv_pos]
-    for instruction in mask_program:
-        op = instruction[0]
-        if op == "const_i32":
-            values.append(fx.Int32(instruction[1]))
-        elif op == "const_bool":
-            constant = fx.Int32(1 if instruction[1] else 0)
-            values.append(constant == fx.Int32(1))
-        elif op == "load_i32":
-            buffer_index = instruction[1]
-            index_ids = instruction[2]
-            offset = fx.Int32(0)
-            for dimension in range_constexpr(len(index_ids)):
-                offset = offset + values[index_ids[dimension]] * fx.Int32(
-                    mask_buffer_strides[buffer_index][dimension]
-                )
-            values.append(load_i32(mask_buffers[buffer_index], offset))
-        else:
-            lhs = values[instruction[1]]
-            if op == "not":
-                values.append(~lhs)
-            else:
-                rhs = values[instruction[2]]
-                if op == "add":
-                    values.append(lhs + rhs)
-                elif op == "sub":
-                    values.append(lhs - rhs)
-                elif op == "mul":
-                    values.append(lhs * rhs)
-                elif op == "floordiv":
-                    values.append(lhs // rhs)
-                elif op == "remainder":
-                    values.append(lhs % rhs)
-                elif op == "ge":
-                    values.append(lhs >= rhs)
-                elif op == "gt":
-                    values.append(lhs > rhs)
-                elif op == "le":
-                    values.append(lhs <= rhs)
-                elif op == "lt":
-                    values.append(lhs < rhs)
-                elif op == "eq":
-                    values.append(lhs == rhs)
-                elif op == "ne":
-                    values.append(lhs != rhs)
-                elif op == "and":
-                    values.append(lhs & rhs)
-                elif op == "or":
-                    values.append(lhs | rhs)
-                else:
-                    raise ValueError(f"unsupported mask bytecode op {op}")
-    return values[mask_program_output]
